+++ src/features/Conversation/index.tsx
@@ -1,13 +1,12 @@
 import { createStyles } from 'antd-style';
-import { ReactNode, memo } from 'react';
+import { ReactNode, Suspense, lazy, memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import ChatHydration from '@/components/StoreHydration/ChatHydration';
-import { useChatStore } from '@/store/chat';
 
 import SkeletonList from './components/SkeletonList';
-import ChatList from './components/VirtualizedList';
-import { useInitConversation } from './hooks/useInitConversation';
+
+const ChatList = lazy(() => import('./components/VirtualizedList'));
 
 const useStyles = createStyles(
   ({ css, responsive, stylish }) => css`
@@ -30,10 +29,6 @@ interface ConversationProps {
 const Conversation = memo<ConversationProps>(({ chatInput, mobile }) => {
   const { styles } = useStyles();
 
-  useInitConversation();
-
-  const [messagesInit] = useChatStore((s) => [s.messagesInit]);
-
   return (
     <Flexbox
       flex={1}
@@ -41,7 +36,9 @@ const Conversation = memo<ConversationProps>(({ chatInput, mobile }) => {
       style={{ position: 'relative' }}
     >
       <div className={styles}>
-        {messagesInit ? <ChatList mobile={mobile} /> : <SkeletonList mobile={mobile} />}
+        <Suspense fallback={<SkeletonList mobile={mobile} />}>
+          <ChatList mobile={mobile} />
+        </Suspense>
       </div>
       {chatInput}
       <ChatHydration />